+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - reorganisation to `sample/test-fns`, `sample/test-rules`
 - [x] `npm run test:rest:all` passes
-- [ ] `npm run test:rules:all` passes
+- [~] `npm run test:rules:all` passes
 
 ## 14-Aug-20